twkevinzhang_mock/requirements.txt
protobuf===5.29.4
grpcio===1.67.1
lxml===5.3.2
aiohttp===3.11.16
beautifulsoup4===4.13.3
